requirements.txt
aerich==0.9.2
aiomysql==0.3.2
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncclick==8.3.0.7
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dictdiffer==0.9.0
fastapi==0.125.0
h11==0.16.0
idna==3.11
iso8601==2.1.0
loguru==0.7.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyMySQL==1.1.2
pypika-tortoise==0.6.3
python-dotenv==1.2.1
pytz==2025.2
starlette==0.50.0
tortoise-orm==0.25.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
